+++ apps/client/src/environments/patch-notes.ts
@@ -1,18 +1,12 @@
 export const patchNotes = `### Bug Fixes
 
-* **community-rotations:** require rlvl before running search.
-* **db:** fix anemone gardens fishing spot name not displayed properly.
-* **db:** fix for instanced fishing holes not loading properly.
-* **lodestone:** fixed lodestone profile sometimes not loading properly.
+* **currency-spending:** fix amount sold last week fixed to 5.
+* **layout:** gathering icon now updates according to its area in breakdown.
+* **lists:** experimental fix for list persistence not applied properly.
+* **simulator:** fix byregot's blessing import for korean rotations.
+* **simulator:** fix level 1-5 recipes being shown as level not met.
 
 
 ### Features
 
-* **db:** support for korean v6.05 update.
-* **desktop:** better packet capture error handling and reporting.
-* **desktop:** support for 6.18 packet capture.
-
-
-### Performance Improvements
-
-* **list:** virtual scrolling buffer for better scrolling behavior.`;
+* **db:** support for korean v6.051 update.`;